fix testing postgis image & remove unused funcs & split layer0 repo
