+++ app/data/repositories/layer0/common.py
@@ -1,0 +1,2 @@
+RAWDATA_SCHEMA = "rawdata"
+INTERNAL_ID_COLUMN_NAME = "hyperleda_internal_id"
