+++ app/data/repositories/layer0/tables.py
@@ -1,21 +1,20 @@
 import datetime
 import json
 
+import pandas
 import structlog
 from astropy import units as u
-from pandas import DataFrame
 
 from app.data import model, template
-from app.lib.storage import enums, postgres
+from app.data.repositories.layer0.common import RAWDATA_SCHEMA
+from app.lib.storage import postgres
 from app.lib.web.errors import DatabaseError
 
-RAWDATA_SCHEMA = "rawdata"
-INTERNAL_ID_COLUMN_NAME = "hyperleda_internal_id"
+log: structlog.stdlib.BoundLogger = structlog.get_logger()
 
 
-class Layer0Repository(postgres.TransactionalPGRepository):
-    def __init__(self, storage: postgres.PgStorage, logger: structlog.stdlib.BoundLogger) -> None:
-        self._logger = logger
+class Layer0TableRepository(postgres.TransactionalPGRepository):
+    def __init__(self, storage: postgres.PgStorage) -> None:
         super().__init__(storage)
 
     def create_table(self, data: model.Layer0TableMeta) -> model.Layer0CreationResponse:
@@ -78,7 +77,7 @@ class Layer0Repository(postgres.TransactionalPGRepository):
         """
 
         if len(data.data) == 0:
-            self._logger.warn("trying to insert 0 rows into the table", table_id=data.table_id)
+            log.warn("trying to insert 0 rows into the table", table_id=data.table_id)
             return
 
         row = self._storage.query_one(template.GET_RAWDATA_TABLE, params=[data.table_id])
@@ -153,7 +152,7 @@ class Layer0Repository(postgres.TransactionalPGRepository):
             params.append(limit)
 
         rows = self._storage.query(query, params=params)
-        return model.Layer0RawData(table_id, DataFrame(rows))
+        return model.Layer0RawData(table_id, pandas.DataFrame(rows))
 
     def fetch_metadata(self, table_id: int) -> model.Layer0TableMeta:
         row = self._storage.query_one(template.GET_RAWDATA_TABLE, params=[table_id])
@@ -218,7 +217,7 @@ class Layer0Repository(postgres.TransactionalPGRepository):
 
         query = template.render_query(
             template.ADD_COLUMN_COMMENT,
-            schema=RAWDATA_SCHEMA,
+            schema="rawdata",
             table_name=table_name,
             column_name=column_description.name,
             params=json.dumps(column_params),
@@ -229,135 +228,10 @@ class Layer0Repository(postgres.TransactionalPGRepository):
             self._storage.exec(query)
             self._storage.exec(modification_query, params=[table_id])
 
-    def upsert_objects(
-        self,
-        table_id: int,
-        objects: list[model.Layer0Object],
-    ) -> None:
-        query = "INSERT INTO rawdata.objects (id, table_id, data) VALUES "
-        params = []
-        values = []
-
-        for obj in objects:
-            values.append("(%s, %s, %s)")
-            params.extend([obj.object_id, table_id, json.dumps(obj.data, cls=model.CatalogObjectEncoder)])
-
-        query += ",".join(values)
-        query += " ON CONFLICT (id) DO UPDATE SET data = EXCLUDED.data, table_id = EXCLUDED.table_id"
-
-        self._storage.exec(query, params=params)
-
-    def upsert_old_object(
-        self,
-        table_id: int,
-        processing_info: model.Layer0OldObject,
-    ) -> None:
-        self._storage.exec(
-            """
-            INSERT INTO rawdata.old_objects (table_id, object_id, status, data, metadata)
-            VALUES (%s, %s, %s, %s, %s) 
-            ON CONFLICT (table_id, object_id) DO 
-                UPDATE SET status = EXCLUDED.status, metadata = EXCLUDED.metadata
-            """,
-            params=[
-                table_id,
-                processing_info.object_id,
-                processing_info.status,
-                json.dumps(processing_info.data, cls=model.CatalogObjectEncoder),
-                json.dumps(processing_info.metadata),
-            ],
-        )
-
-    def get_table_statistics(self, table_id: int) -> model.TableStatistics:
-        statuses_query = """
-            SELECT COALESCE(status, 'unprocessed') AS status, COUNT(1) 
-            FROM rawdata.processing AS p
-            RIGHT JOIN rawdata.objects AS o ON p.object_id = o.id
-            WHERE o.table_id = %s
-            GROUP BY status"""
-
-        stats_query = """
-            SELECT MAX(modification_dt) AS modification_dt , COUNT(1) AS cnt
-            FROM rawdata.objects 
-            WHERE table_id = %s"""
-
-        status_rows = self._storage.query(statuses_query, params=[table_id])
-        stats = self._storage.query_one(stats_query, params=[table_id])
-
-        table_name = self._storage.query_one(template.GET_RAWDATA_TABLE, params=[table_id]).get("table_name")
-        if table_name is None:
-            raise DatabaseError(f"unable to fetch table with id {table_id}")
-
-        total_original_rows_query = f'SELECT COUNT(1) AS cnt FROM {RAWDATA_SCHEMA}."{table_name}"'
-
-        total_original_rows = self._storage.query_one(total_original_rows_query).get("cnt")
-        if total_original_rows is None:
-            raise DatabaseError(f"unable to fetch total rows for table {table_name}")
-
-        return model.TableStatistics(
-            {enums.ObjectProcessingStatus(row["status"]): row["count"] for row in status_rows},
-            stats["modification_dt"],
-            stats["cnt"],
-            total_original_rows,
-        )
-
-    def get_old_object_statuses(self, table_id: int) -> dict[enums.ObjectProcessingStatus, int]:
-        rows = self._storage.query(
-            """
-            SELECT status, COUNT(*) FROM rawdata.old_objects 
-            WHERE table_id = %s 
-            GROUP BY status
-            """,
-            params=[table_id],
-        )
-
-        return {enums.ObjectProcessingStatus(row["status"]): row["count"] for row in rows}
-
-    def get_objects(self, table_id: int, limit: int, offset: int) -> list[model.Layer0Object]:
-        rows = self._storage.query(
-            """
-            SELECT id, data
-            FROM rawdata.objects
-            WHERE table_id = %s
-            LIMIT %s OFFSET %s
-            """,
-            params=[table_id, limit, offset],
-        )
-
-        return [
-            model.Layer0Object(
-                row["id"],
-                json.loads(json.dumps(row["data"]), cls=model.CatalogObjectDecoder),
-            )
-            for row in rows
-        ]
-
-    def get_old_objects(self, table_id: int, batch_size: int, offset: int) -> list[model.Layer0OldObject]:
-        rows = self._storage.query(
-            """
-            SELECT object_id, pgc, status, data, metadata
-            FROM rawdata.old_objects
-            WHERE table_id = %s
-            LIMIT %s OFFSET %s
-            """,
-            params=[table_id, batch_size, offset],
-        )
-
-        return [
-            model.Layer0OldObject(
-                row["object_id"],
-                enums.ObjectProcessingStatus(row["status"]),
-                row["metadata"],
-                json.loads(json.dumps(row["data"]), cls=model.CatalogObjectDecoder),
-                row["pgc"],
-            )
-            for row in rows
-        ]
-
     def _get_table_id(self, table_name: str) -> tuple[int, bool]:
         try:
             row = self._storage.query_one(
-                f"SELECT id FROM {RAWDATA_SCHEMA}.tables WHERE table_name = %s",
+                "SELECT id FROM rawdata.tables WHERE table_name = %s",
                 params=[table_name],
             )
         except RuntimeError:
