+++ tests/lib/postgres.py
@@ -26,7 +26,7 @@ class TestPostgresStorage:
         self.port = web.find_free_port()
         logger.info("Initializing postgres container", port=self.port)
         self.container = pgcontainer.PostgresContainer(
-            "postgis/postgis:17-3.5@sha256:4a962e61b8cf3543d920859dd6583d8248767f89aea9e03ad5ec715c735665c2",
+            "postgis/postgis:17-3.5-alpine@sha256:f63cf3c8acfd305a5b33d34b2667509b53465924999fe2ec276166080c16319b",
             port=5432,
             user="hyperleda",
             password="password",
